requirements.txt
streamlit
requests
pymongo

bcrypt
pymodm
bson

python-dotenv